refactor(alignment): 人工覆盖 CRUD 外移，text_alignment 2855→2484
覆盖（提案/确认/撤销/查询）不是对齐核心的算法，是核心之上的人工干预层：
它消费定位与校验能力，核心不反向依赖它。闭包分析确认零回边，故往上拆一层。

  alignment_overrides  412 行  resolve_override_context / create_override_proposal
                              / confirm_override / revoke_override / list_overrides

新模块在上层，无法像上一轮那样从 text_alignment 重新导出（会成环），故按实际
消费者改导入：application/alignment_override_service 的 4 处懒 import 与
tests/test_alignment_overrides。text_alignment.py 的改动是纯删除，未改写函数体。

配方快照（read/restore/replace_alignment_recipe_snapshot，111 行）同属这一层且
与覆盖零耦合，本轮**未**拆出：database 的重建路径要 import 它，移出会把
test_no_new_import_cycles_appear 冻结的已知 4 节点环
（database ↔ text_alignment ↔ calibration_library ↔ bibliographic_metadata）
拉长为 5 节点。已实测确认，故退回。待 bibliographic_metadata 解开对 database
的领域纠缠后一并处理。

行数上限只降不升：text_alignment 2875→2500，新增 alignment_overrides 425。

全量 unittest discover 1981 项通过（1 跳过）；导入环守卫绿，已知环节点数未增。
